requirements.txt
flask
flask-migrate
python-dotenv
flask-wtf
email-validator